+++ frontend/lib/presentation/screens/main/downloaded/small_widget_download.dart
@@ -1,9 +1,7 @@
 
 
 import 'package:flutter/material.dart';
-import 'package:frontend/presentation/boilerplate/navigation_bar.dart';
-
-// import 'package:frontend/presentation/assets/l10n/generated/l10n.dart';
+import 'package:frontend/presentation/assets/l10n/generated/l10n.dart';
 
 
 
@@ -11,18 +9,21 @@ import 'package:frontend/presentation/boilerplate/navigation_bar.dart';
 
 
 
-class SmallWidgetSearch extends StatefulWidget {
+
+
+
+class SmallWidgetDownload extends StatefulWidget {
   final Function(Locale) onLocaleChange;
 
-  const SmallWidgetSearch({
+  const SmallWidgetDownload({
     super.key, required this.onLocaleChange
     });
 
   @override
-  SmallWidgetSearchState createState() => SmallWidgetSearchState();
+  SmallWidgetDownloadState createState() => SmallWidgetDownloadState();
 }
 
-class SmallWidgetSearchState extends State<SmallWidgetSearch> {
+class SmallWidgetDownloadState extends State<SmallWidgetDownload> {
 
   
 
@@ -30,31 +31,8 @@ class SmallWidgetSearchState extends State<SmallWidgetSearch> {
   @override
   Widget build(BuildContext context) {
 
-    
-    
-    return Stack(
-        children: [
-          // Gradient background
-            AnimatedContainer(
-              
-              duration: Duration(seconds: 3),
-              decoration: BoxDecoration(
-                gradient: LinearGradient(
-                  begin: Alignment.topCenter,
-                  end: Alignment.bottomCenter,
-                  colors: [
-                  Color(0xFFD6A49B).withOpacity(0.7), // Soft pink
-                  Colors.grey.shade300,
-                  Colors.white.withOpacity(1), // Return to soft pink
-                            ],
-                ),
-              ),
-            ), Scaffold(
-             extendBodyBehindAppBar: true, // Allows content to extend behind the AppBar
-            backgroundColor: Colors.transparent,
-       
-        bottomNavigationBar: MyNavigationBar(),
-        body: SingleChildScrollView(
+  
+    return SingleChildScrollView(
           child: Column( 
             mainAxisAlignment: MainAxisAlignment.spaceBetween,
             crossAxisAlignment: CrossAxisAlignment.center,
@@ -69,7 +47,7 @@ class SmallWidgetSearchState extends State<SmallWidgetSearch> {
                 width: 400,
               child: TextField(
                 decoration: InputDecoration(
-                  hintText: "Search here...",
+                  hintText: S.of(context).searchBarPhrase,
                   filled: true,
                   fillColor: Colors.white.withOpacity(0.8),
                   prefixIcon: Icon(Icons.search),
@@ -84,8 +62,7 @@ class SmallWidgetSearchState extends State<SmallWidgetSearch> {
             ElevatedButton.icon(
                         onPressed: () {
                         },
-                        label: Text(
-                          "Search",
+                        label: Text(S.of(context).search,
                           style: TextStyle(color: Theme.of(context).colorScheme.onSecondary),
                         ),
                         style: ElevatedButton.styleFrom(
@@ -107,8 +84,8 @@ class SmallWidgetSearchState extends State<SmallWidgetSearch> {
            Spacer(),
               ])
           
-          ])),
-      )]);
+          ]));
+
 
   }